added bounce off of screen edge to virus (see description)
Added impulse to virus spawn, not working in the correct screen bounds yet, but still bouncing around the screen. Found that virus 2 moves a lot faster straight down than virus 1. Unsure as to why I have this bug however, it provides better differentiation between the two virus types so I am going to keep it as a feature.